+++ earthdaily/earthdatastore/cube_utils/__init__.py
@@ -3,7 +3,8 @@ import numpy as np
 import pandas as pd
 import geopandas as gpd
 from shapely.geometry import box
-from earthdaily.earthdatastore.cube_utils._zonal import zonal_stats, zonal_stats_numpy
+from ._zonal import zonal_stats, zonal_stats_numpy
+from .asset_mapper import AssetMapper
 from rasterio.enums import Resampling
 from rasterio.mask import geometry_mask
 import rioxarray as rxr
@@ -42,7 +43,7 @@ def _cube_odc(items_collection, assets=None, times=None, **kwargs):
     if "resampling" in kwargs:
         if isinstance(kwargs["resampling"], int):
             kwargs["resampling"] = Resampling(kwargs["resampling"]).name
-    chunks = kwargs.get("chunks", dict(x=2048, y=2048, time=1))
+    chunks = kwargs.get("chunks", dict(x="auto", y="auto", time="auto"))
     kwargs.pop("chunks", None)
 
     ds = stac.load(
@@ -90,6 +91,7 @@ def datacube(
     engine="odc",
     rescale=True,
     groupby_date="mean",
+    common_band_names=True,
     **kwargs,
 ):
     logging.info(f"Building datacube with {len(items_collection)} items")
@@ -106,6 +108,10 @@ def datacube(
         raise NotImplementedError(
             f"Engine '{engine}' not supported. Only {' and '.join(list(engines.keys()))} are currently supported."
         )
+    if common_band_names and not isinstance(assets, dict):
+        aM = AssetMapper()
+        assets = aM.map_collection_bands(items_collection[0].collection_id, assets)
+
     if isinstance(assets, dict):
         assets_keys = list(assets.keys())
     ds = engines[engine](
@@ -192,7 +198,7 @@ def rescale_assets_with_items(
     scales = dict()
     if len(items_collection) > ds.time.size:
         unique_dt = {}
-        items_collection_unique_dt = []
+        # items_collection_unique_dt = []
         for item in items_collection:
             if item.datetime in unique_dt.keys():
                 for asset in item.assets.keys():
@@ -245,7 +251,7 @@ def rescale_assets_with_items(
             scale = rasterbands.get("scale", None)
 
             if offset or scale:
-                if not ds_asset in scales:
+                if ds_asset not in scales:
                     scales[ds_asset] = {}
 
                 scale = rasterbands.get("scale", 1)