+++ earthdaily/earthdatastore/cube_utils/_zonal.py
@@ -16,9 +16,7 @@ from .preprocessing import rasterize
 
 def _compute_M(data):
     cols = np.arange(data.size)
-    return csr_matrix(
-        (cols, (data.ravel(), cols)), shape=(data.max() + 1, data.size)
-    )
+    return csr_matrix((cols, (data.ravel(), cols)), shape=(data.max() + 1, data.size))
 
 
 def _indices_sparse(data):
@@ -45,11 +43,8 @@ def datacube_time_stats(datacube, operations):
 
 def _rasterize(gdf, dataset, all_touched=False):
     feats = rasterize(gdf, dataset, all_touched=all_touched)
-    idx_start = 0
-    if 0 in feats:
-        idx_start = 1
     yx_pos = _indices_sparse(feats)
-    return feats, yx_pos, idx_start
+    return feats, yx_pos
 
 
 def zonal_stats_numpy(
@@ -60,11 +55,11 @@ def zonal_stats_numpy(
     preload_datavar=False,
 ):
     tqdm_bar = tqdm.tqdm(total=len(dataset.data_vars) * dataset.time.size)
+    dataset = dataset.rio.clip_box(*gdf.to_crs(dataset.rio.crs).total_bounds)
 
-    feats, yx_pos, idx_start = _rasterize(
-        gdf, dataset, all_touched=all_touched
-    )
+    feats, yx_pos = _rasterize(gdf, dataset, all_touched=all_touched)
     ds = []
+    features_idx = []
     for data_var in dataset.data_vars:
         tqdm_bar.set_description(data_var)
         dataset_var = dataset[data_var]
@@ -76,9 +71,19 @@ def zonal_stats_numpy(
             vals[t] = []
             mem_asset = dataset_var.isel(time=t).to_numpy()
             for i in range(gdf.shape[0]):
-                pos = yx_pos[i + idx_start]
-                data = mem_asset[pos]
-                res = [operation(data) for operation in operations.values()]
+                features_idx.append(i)
+                if len(yx_pos) <= i + 1:
+                    break
+                pos = np.asarray(yx_pos[i + 1])
+                # mem_asset[*pos] only for python>=3.11
+                if len(pos) == 2:
+                    data = mem_asset[pos[0], pos[1]]
+                elif len(pos) == 1:
+                    data = mem_asset[pos[0]]
+                if data.size > 0:
+                    res = [operation(data) for operation in operations.values()]
+                else:
+                    res = [np.nan for operation in operations]
                 vals[t].append(res)
         arr = np.asarray([vals[v] for v in vals])
 
@@ -87,14 +92,14 @@ def zonal_stats_numpy(
             dims=["time", "feature", "stats"],
             coords=dict(
                 time=dataset_var.time.values,
-                feature=gdf.index,
+                feature=gdf.index[np.nonzero(np.unique(feats))[0] - 1],
                 stats=list(operations.keys()),
             ),
         )
         del arr, mem_asset, vals, dataset_var
         ds.append(da.to_dataset(name=data_var))
     tqdm_bar.close()
-    return xr.merge(ds)
+    return xr.merge(ds).transpose("feature", "time", "stats")
 
 
 def zonal_stats(
@@ -115,23 +120,27 @@ def zonal_stats(
 
     zonal_ds_list = []
 
+    dataset = dataset.rio.clip_box(*gdf.to_crs(dataset.rio.crs).total_bounds)
+
     if method == "optimized":
-        feats, yx_pos, idx_start = _rasterize(
-            gdf, dataset, all_touched=all_touched
-        )
+        feats, yx_pos = _rasterize(gdf, dataset, all_touched=all_touched)
 
         for gdf_idx in tqdm.trange(gdf.shape[0], disable=not verbose):
             tqdm_bar.update(1)
-            yx_pos_idx = yx_pos[gdf_idx + idx_start]
+            if gdf_idx + 1 >= len(yx_pos):
+                continue
+            yx_pos_idx = yx_pos[gdf_idx + 1]
+            if np.asarray(yx_pos_idx).size == 0:
+                continue
             datacube_spatial_subset = dataset.isel(
                 x=xr.DataArray(yx_pos_idx[1], dims="xy"),
                 y=xr.DataArray(yx_pos_idx[0], dims="xy"),
             )
             del yx_pos_idx
             zonal_ds_list.append(
-                datacube_time_stats(
-                    datacube_spatial_subset, operations
-                ).expand_dims(dim={"feature": [gdf.iloc[gdf_idx].name]})
+                datacube_time_stats(datacube_spatial_subset, operations).expand_dims(
+                    dim={"feature": [gdf.iloc[gdf_idx].name]}
+                )
             )
 
         del yx_pos, feats
@@ -145,9 +154,7 @@ def zonal_stats(
                 shapes = feat.geometry.geoms
             else:
                 shapes = [feat.geometry]
-            datacube_spatial_subset = dataset.rio.clip(
-                shapes, all_touched=all_touched
-            )
+            datacube_spatial_subset = dataset.rio.clip(shapes, all_touched=all_touched)
 
             zonal_feat = datacube_time_stats(
                 datacube_spatial_subset, operations
@@ -155,7 +162,5 @@ def zonal_stats(
 
             zonal_ds_list.append(zonal_feat)
     else:
-        raise NotImplementedError(
-            'method available are : "standard" or "optimized"'
-        )
+        raise NotImplementedError('method available are : "standard" or "optimized"')
     return xr.concat(zonal_ds_list, dim="feature")