+++ earthdaily/earthdatastore/cube_utils/asset_mapper/_asset_mapper_config.py
@@ -1,0 +1,158 @@
+asset_mapper_collections = {
+    "sentinel-2-l2a": [
+        {
+            "coastal44": "coastal",
+            "coastal": "coastal",
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+            "nir08": "nir08",
+            "nir09": "nir09",
+            "swir16": "swir16",
+            "swir22": "swir22",
+            "rededge": "rededge2",
+            "rededge70": "rededge1",
+            "rededge74": "rededge2",
+            "rededge78": "rededge3",
+            "watervapor": "wvp",
+            "native-cloud-mask": "scl",
+        }
+    ],
+    "landsat-c2l2-sr": [
+        {
+            "coastal44": "coastal",
+            "coastal": "coastal",
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir08",
+            "nir08": "nir08",
+            "swir16": "swir16",
+            "swir22": "swir22",
+            "native-cloud-mask": "qa-pixel",
+        }
+    ],
+    "landsat-c2l2-st": [
+        {"lwir": "lwir11", "lwir11": "lwir11", "native-cloud-mask": "qa-pixel"}
+    ],
+    "venus-l2a": [
+        {
+            "coastal42": "image_file_SRE_B1",
+            "coastal44": "image_file_SRE_B2",
+            "coastal": "image_file_SRE_B2",
+            "blue": "image_file_SRE_B3",
+            "green": "image_file_SRE_B4",
+            "yellow": "image_file_SRE_B5",
+            "red": "image_file_SRE_B7",
+            "nir": "image_file_SRE_B11",
+            "nir08": "image_file_SRE_B11",
+            "nir09": "image_file_SRE_B12",
+            "rededge": "image_file_SRE_B09",
+            "rededge70": "image_file_SRE_B08",
+            "rededge74": "image_file_SRE_B09",
+            "rededge78": "image_file_SRE_B10",
+            "native-cloud-mask": "detailed_cloud_mask",
+        }
+    ],
+    "earthdaily-simulated-cloudless-l2a-cog-edagro": [
+        {
+            "coastal44": "image_file_C",
+            "coastal": "image_file_C",
+            "blue": "image_file_B",
+            "green": "image_file_G",
+            "yellow": "image_file_Y",
+            "red": "image_file_R",
+            "nir": "image_file_NIR",
+            "nir08": "image_file_NIR",
+            "swir12": "image_file_SWIR1",
+            "swir16": "image_file_SWIR2",
+            "cirrus": "image_file_CI",
+            "rededge": "image_file_RE2",
+            "rededge70": "image_file_RE1",
+            "rededge74": "image_file_RE2",
+            "rededge78": "image_file_RE3",
+            "watervapor1": "image_file_RWVP1",
+            "watervapor2": "image_file_RWVP2",
+            "lwir": "image_file_ST",
+            "lwir11": "image_file_ST",
+            "tir1": "image_file_TIR",
+            "tir2": "image_file_TIR2",
+            "tir3": "image_file_TIR3",
+            "aqua": "image_file_A",
+            "wildfire": "image_file_W",
+            "me1": "image_file_ME1",
+            "me2": "image_file_ME2",
+        }
+    ],
+    "alsat-1b-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "cbers-4-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "deimos-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "gaofen-1-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "gaofen-6-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "kazstsat-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "huanjing-2-l2a-cog-edagro": [
+        {
+            "blue": "blue",
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "dmc-l2a-cog-edagro": [
+        {
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+    "resourcesat-2-l2a-cog-edagro": [
+        {
+            "green": "green",
+            "red": "red",
+            "nir": "nir",
+        }
+    ],
+}