+++ earthdaily/earthdatastore/mask/__init__.py
@@ -4,16 +4,11 @@ from rasterio.features import geometry_mask
 from earthdaily.earthdatastore.cube_utils import _bbox_to_intersects
 import geopandas as gpd
 import warnings
-import json
 import numpy as np
 import tqdm
 from joblib import Parallel, delayed
 
 dask.config.set(**{"array.slicing.split_large_chunks": True})
-
-warnings.simplefilter(
-    "ignore", category=xr.core.extensions.AccessorRegistrationWarning
-)
 
 _available_masks = [
     "native",
@@ -35,6 +30,17 @@ _native_mask_asset_mapping = {
 }
 
 
+def _bool_or_int_to_njobs(var):
+    if isinstance(var, bool):
+        if var:
+            arg = 1
+        else:
+            arg = False
+    else:
+        arg = var
+    return arg
+
+
 class Mask:
     def __init__(self, dataset: xr.Dataset, intersects=None, bbox=None):
         self._obj = dataset
@@ -50,25 +56,19 @@ class Mask:
         add_mask_var=False,
         mask_statistics=False,
     ):
-        acm_datacube["time"] = acm_datacube.time.dt.round(
-            "s"
-        )  # rm nano second
+        acm_datacube["time"] = acm_datacube.time.dt.round("s")  # rm nano second
         self._obj["time"] = self._obj.time.dt.round("s")  # rm nano second
         #
-        self._obj = self._obj.where(
-            acm_datacube["agriculture-cloud-mask"] == 1
-        )
+        self._obj = self._obj.where(acm_datacube["agriculture-cloud-mask"] == 1)
         if add_mask_var:
-            self._obj["agriculture-cloud-mask"] = acm_datacube[
-                "agriculture-cloud-mask"
-            ]
+            self._obj["agriculture-cloud-mask"] = acm_datacube["agriculture-cloud-mask"]
         if mask_statistics:
             self.compute_clear_coverage(
                 acm_datacube["agriculture-cloud-mask"],
                 "ag_cloud_mask",
                 1,
                 labels_are_clouds=False,
-                n_jobs=1 if mask_statistics == True else mask_statistics,
+                n_jobs=_bool_or_int_to_njobs(mask_statistics),
             )
         return self._obj
 
@@ -82,9 +82,7 @@ class Mask:
         fill_value=np.nan,
     ):
         if cloud_asset not in self._obj.data_vars:
-            raise ValueError(
-                f"Asset '{cloud_asset}' needed to compute cloudmask."
-            )
+            raise ValueError(f"Asset '{cloud_asset}' needed to compute cloudmask.")
         else:
             cloud_layer = self._obj[cloud_asset].copy()
         _assets = [a for a in self._obj.data_vars if a != cloud_asset]
@@ -105,7 +103,7 @@ class Mask:
                 cloud_asset,
                 labels,
                 labels_are_clouds=labels_are_clouds,
-                n_jobs=1 if mask_statistics == True else mask_statistics,
+                n_jobs=_bool_or_int_to_njobs(mask_statistics),
             )
         return self._obj
 
@@ -123,9 +121,7 @@ class Mask:
             mask_statistics=mask_statistics,
         )
 
-    def venus_detailed_cloud_mask(
-        self, add_mask_var=False, mask_statistics=False
-    ):
+    def venus_detailed_cloud_mask(self, add_mask_var=False, mask_statistics=False):
         return self.cloudmask_from_asset(
             "detailed_cloud_mask",
             0,
@@ -142,15 +138,13 @@ class Mask:
         labels_are_clouds=True,
         n_jobs=1,
     ):
-        def compute_clear_pixels(
-            cloudmask_array, labels, labels_are_clouds=False
-        ):
+        def compute_clear_pixels(cloudmask_array, labels, labels_are_clouds=False):
             cloudmask_array = cloudmask_array.data.compute()
 
             if labels_are_clouds:
-                labels_sum = np.sum(
-                    ~np.in1d(cloudmask_array, labels)
-                ) - np.sum(np.isnan(cloudmask_array))
+                labels_sum = np.sum(~np.in1d(cloudmask_array, labels)) - np.sum(
+                    np.isnan(cloudmask_array)
+                )
             else:
                 labels_sum = np.sum(np.in1d(cloudmask_array, labels))
             return labels_sum
@@ -173,6 +167,10 @@ class Mask:
         self._obj = self._obj.assign_coords(
             {f"clear_pixels_{cloudmask_name}": ("time", n_pixels_as_labels)}
         )
+        self._obj.coords["clear_pixels"] = self._obj.coords[
+            f"clear_pixels_{cloudmask_name}"
+        ]
+
         self._obj = self._obj.assign_coords(
             {
                 f"clear_percent_{cloudmask_name}": (
@@ -184,6 +182,15 @@ class Mask:
                 )
             }
         )
+
+        self._obj.coords["clear_percent"] = self._obj.coords[
+            f"clear_percent_{cloudmask_name}"
+        ]
+        warnings.warn(
+            f"Removed in 0.0.2 : The two coordinates 'clear_pixels_{cloudmask_name}' and 'clear_percent_{cloudmask_name}' will be deleted to keep only the prefix 'clear_percent' and 'clear_pixels'",
+            category=DeprecationWarning,
+        )
+
         return self._obj
 
     def compute_available_pixels(self):
@@ -253,3 +260,7 @@ def QA_PIXEL_cloud_detection(arr):
     cloudfree_pixels = cloudfree[cloudfree != 0]
     cloudmask = np.isin(arr, cloudfree_pixels).astype(int)
     return cloudmask
+
+
+def filter_clear_cover(dataset, clear_cover, coordinate="clear_percent"):
+    return dataset.sel(time=dataset.time[dataset.clear_percent >= clear_cover])