+++ examples/summary_stack.py
@@ -7,11 +7,9 @@ Explore collection metadata using earthdaily
 # Import librairies
 # -------------------------------------------
 
-import json
 from earthdaily import earthdatastore
 from rich.table import Table
 from rich.console import Console
-import os
 
 from dotenv import load_dotenv
 
@@ -40,7 +38,7 @@ console.print(table)
 # Explore a specific collection
 # -------------------------------------------
 
-collection = eds.explore("sentinel-2-l2a")
+collection = eds.explore("venus-l2a")
 console.log(collection.properties)
 
 ##############################################################################